requirements.txt
Django==2.0.1
djangorestframework==3.7.7
gunicorn==19.7.1
pytz==2017.3
